refactor(ui): déplacer le menu utilisateur dans le bas de la sidebar
- Déplace le DropdownMenu (email, paramètres, langue, thème, déconnexion)
  du header vers le footer de la sidebar desktop
- Ajoute l'option "Code source" (lien GitHub) dans le menu
- Ajoute le numéro de version (branch@commit) comme libellé non-cliquable
- Simplifie le header : conserve uniquement le trigger mobile (Sheet)
