+++ client/src/components/layout/sidebar.tsx
@@ -1,28 +1,68 @@
 "use client";
 
 import Link from "next/link";
-import { Github, MoreHorizontal, Plus } from "lucide-react";
+import { signOut } from "next-auth/react";
+import { Check, Github, Languages, LogOut, Monitor, Moon, MoreHorizontal, Plus, Settings, Sun } from "lucide-react";
 import { usePathname } from "next/navigation";
 import { useQueries } from "@tanstack/react-query";
-import { useTranslations } from "next-intl";
+import { useLocale, useTranslations } from "next-intl";
 import { Button } from "@/components/ui/button";
 import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
+  DropdownMenuSeparator,
+  DropdownMenuSub,
+  DropdownMenuSubContent,
+  DropdownMenuSubTrigger,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu";
 import { useProjects } from "@/lib/hooks/use-projects";
 import { useAuth } from "@/lib/hooks/use-auth";
 import { listOrganizationMembers, listOrganizationProjects } from "@/lib/api/organizations";
 import { useOrganizations } from "@/lib/hooks/use-organizations";
+import type { Theme } from "@/lib/providers/theme-provider";
+import { useTheme } from "@/lib/providers/theme-provider";
+import { useUpdateUserLocale } from "@/lib/hooks/use-user-profile";
+import type { UpdateUserLocaleRequest } from "@/lib/types/api";
 import { cn } from "@/lib/utils";
+
+const LOCALES: Array<{ value: UpdateUserLocaleRequest["locale"]; label: string }> = [
+  { value: "en", label: "English" },
+  { value: "fr", label: "Français" },
+];
+
+function setLocaleCookie(locale: string) {
+  const maxAge = 60 * 60 * 24 * 365;
+  document.cookie = `raggae_locale=${locale};path=/;max-age=${maxAge};SameSite=Lax`;
+}
 
 export function Sidebar() {
   const pathname = usePathname();
   const t = useTranslations("sidebar");
+  const tHeader = useTranslations("header");
   const tNav = useTranslations("nav");
   const tCommon = useTranslations("common");
+  const tTheme = useTranslations("layout.themeToggle");
+  const { theme, setTheme } = useTheme();
+  const currentLocale = useLocale();
+  const updateLocale = useUpdateUserLocale();
+
+  const THEMES: Array<{ value: Theme; label: string; icon: React.ReactNode }> = [
+    { value: "light", label: tTheme("light"), icon: <Sun size={14} /> },
+    { value: "dark", label: tTheme("dark"), icon: <Moon size={14} /> },
+    { value: "system", label: tTheme("system"), icon: <Monitor size={14} /> },
+  ];
+
+  function handleThemeSelect(next: Theme) {
+    if (next !== theme) setTheme(next);
+  }
+
+  function handleLocaleSelect(locale: UpdateUserLocaleRequest["locale"]) {
+    if (locale === currentLocale) return;
+    setLocaleCookie(locale);
+    updateLocale.mutate({ locale }, { onSettled: () => window.location.reload() });
+  }
 
   const navItems = [
     { href: "/projects", label: tNav("projects"), icon: "folder" },
@@ -231,18 +271,100 @@ export function Sidebar() {
         ))}
       </nav>
       <div className="border-t p-3">
-        <a
-          href="https://github.com/jbuget/raggae"
-          target="_blank"
-          rel="noreferrer"
-          className="flex items-center gap-2 text-sm text-muted-foreground hover:text-foreground"
-        >
-          <Github className="size-4" />
-          <span>{t("githubProject")}</span>
-        </a>
-        <p className="mt-2 text-xs text-muted-foreground">
-          {gitBranch}@{shortCommit}
-        </p>
+        <DropdownMenu>
+          <DropdownMenuTrigger asChild>
+            <Button variant="ghost" size="sm" className="w-full cursor-pointer justify-start truncate">
+              {user?.email ?? tHeader("account")}
+            </Button>
+          </DropdownMenuTrigger>
+          <DropdownMenuContent side="top" align="start" className="w-52">
+            <DropdownMenuItem disabled className="text-xs text-muted-foreground">
+              {user?.email}
+            </DropdownMenuItem>
+
+            <DropdownMenuSeparator />
+
+            <DropdownMenuItem asChild>
+              <Link href="/settings" className="flex cursor-pointer items-center gap-2">
+                <Settings size={14} />
+                {tHeader("userSettings")}
+              </Link>
+            </DropdownMenuItem>
+
+            <DropdownMenuSub>
+              <DropdownMenuSubTrigger className="flex cursor-pointer items-center gap-2">
+                <Languages size={14} />
+                {tHeader("language")}
+              </DropdownMenuSubTrigger>
+              <DropdownMenuSubContent>
+                {LOCALES.map(({ value, label }) => (
+                  <DropdownMenuItem
+                    key={value}
+                    onClick={() => handleLocaleSelect(value)}
+                    className="flex cursor-pointer items-center gap-2"
+                  >
+                    <Check
+                      size={14}
+                      className={currentLocale === value ? "opacity-100" : "opacity-0"}
+                    />
+                    {label}
+                  </DropdownMenuItem>
+                ))}
+              </DropdownMenuSubContent>
+            </DropdownMenuSub>
+
+            <DropdownMenuSub>
+              <DropdownMenuSubTrigger className="flex cursor-pointer items-center gap-2">
+                <Sun size={14} />
+                {tTheme("label")}
+              </DropdownMenuSubTrigger>
+              <DropdownMenuSubContent>
+                {THEMES.map(({ value, label, icon }) => (
+                  <DropdownMenuItem
+                    key={value}
+                    onClick={() => handleThemeSelect(value)}
+                    className="flex cursor-pointer items-center gap-2"
+                  >
+                    <Check
+                      size={14}
+                      className={theme === value ? "opacity-100" : "opacity-0"}
+                    />
+                    {icon}
+                    {label}
+                  </DropdownMenuItem>
+                ))}
+              </DropdownMenuSubContent>
+            </DropdownMenuSub>
+
+            <DropdownMenuSeparator />
+
+            <DropdownMenuItem asChild>
+              <a
+                href="https://github.com/jbuget/raggae"
+                target="_blank"
+                rel="noreferrer"
+                className="flex cursor-pointer items-center gap-2"
+              >
+                <Github size={14} />
+                {t("sourceCode")}
+              </a>
+            </DropdownMenuItem>
+
+            <DropdownMenuItem disabled className="text-xs text-muted-foreground">
+              {gitBranch}@{shortCommit}
+            </DropdownMenuItem>
+
+            <DropdownMenuSeparator />
+
+            <DropdownMenuItem
+              onClick={() => signOut({ callbackUrl: "/login" })}
+              className="flex cursor-pointer items-center gap-2"
+            >
+              <LogOut size={14} />
+              {tHeader("signOut")}
+            </DropdownMenuItem>
+          </DropdownMenuContent>
+        </DropdownMenu>
       </div>
     </aside>
   );
